+++ forui/lib/src/widgets/alert.dart
@@ -15,7 +15,7 @@ import 'package:forui/forui.dart';
 class FAlert extends StatelessWidget {
   /// The icon. Defaults to [FAssets.icons.circleAlert].
   ///
-  /// Icons are wrapped in [FIconData], and therefore works with [FIcon]s.
+  /// [icon] is wrapped in [FIconStyle], and therefore works with [FIcon]s.
   final Widget icon;
 
   /// The title.
@@ -63,9 +63,8 @@ class FAlert extends StatelessWidget {
           children: [
             Row(
               children: [
-                FIconData(
-                  color: style.iconColor,
-                  size: style.iconSize,
+                FInheritedIconStyle(
+                  style: FIconStyle(color: style.iconColor, size: style.iconSize),
                   child: icon,
                 ),
                 Flexible(
@@ -105,6 +104,84 @@ class FAlert extends StatelessWidget {
     super.debugFillProperties(properties);
     properties.add(DiagnosticsProperty('style', style));
   }
+}
+
+/// [FAlertCustomStyle]'s style.
+final class FAlertStyles with Diagnosticable {
+  /// The primary alert style.
+  final FAlertCustomStyle primary;
+
+  /// The destructive alert style.
+  final FAlertCustomStyle destructive;
+
+  /// Creates a [FAlertStyles].
+  const FAlertStyles({
+    required this.primary,
+    required this.destructive,
+  });
+
+  /// Creates a [FAlertStyles] that inherits its properties from the provided [colorScheme], [typography], and [style].
+  FAlertStyles.inherit({required FColorScheme colorScheme, required FTypography typography, required FStyle style})
+      : primary = FAlertCustomStyle(
+    padding: const EdgeInsets.fromLTRB(16, 12, 16, 12),
+    iconColor: colorScheme.foreground,
+    titleTextStyle: typography.base.copyWith(
+      fontWeight: FontWeight.w500,
+      color: colorScheme.foreground,
+      height: 1.2,
+    ),
+    subtitleTextStyle: typography.sm.copyWith(color: colorScheme.foreground),
+    decoration: BoxDecoration(
+      border: Border.all(color: colorScheme.border),
+      borderRadius: style.borderRadius,
+      color: colorScheme.background,
+    ),
+  ),
+        destructive = FAlertCustomStyle(
+          padding: const EdgeInsets.fromLTRB(16, 12, 16, 12),
+          iconColor: colorScheme.destructive,
+          titleTextStyle: typography.base.copyWith(
+            fontWeight: FontWeight.w500,
+            color: colorScheme.destructive,
+            height: 1.2,
+          ),
+          subtitleTextStyle: typography.sm.copyWith(color: colorScheme.destructive),
+          decoration: BoxDecoration(
+            border: Border.all(color: colorScheme.destructive),
+            borderRadius: style.borderRadius,
+            color: colorScheme.background,
+          ),
+        );
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty('primary', primary))
+      ..add(DiagnosticsProperty('destructive', destructive));
+  }
+
+  /// Returns a copy of this [FAlertStyles] with the given properties replaced.
+  @useResult
+  FAlertStyles copyWith({
+    FAlertCustomStyle? primary,
+    FAlertCustomStyle? destructive,
+  }) =>
+      FAlertStyles(
+        primary: primary ?? this.primary,
+        destructive: destructive ?? this.destructive,
+      );
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+          other is FAlertStyles &&
+              runtimeType == other.runtimeType &&
+              primary == other.primary &&
+              destructive == other.destructive;
+
+  @override
+  int get hashCode => primary.hashCode ^ destructive.hashCode;
 }
 
 /// A [FAlert]'s style.