+++ forui/lib/src/widgets/icon.dart
@@ -1,11 +1,35 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
-import 'package:forui/assets.dart';
+import 'package:flutter/material.dart';
+import 'package:forui/forui.dart';
+
+/// The [FIconStyle] that this [FInheritedIconStyle]'s widget subtree should inherit.
+class FInheritedIconStyle extends InheritedWidget {
+  /// The icon's data.
+  final FIconStyle style;
+
+  /// Creates a [FIconStyle].
+  const FInheritedIconStyle({
+    required this.style,
+    required super.child,
+    super.key,
+  });
+
+  @override
+  bool updateShouldNotify(FInheritedIconStyle old) => style != old.style;
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(DiagnosticsProperty('style', style));
+  }
+}
 
 /// The default properties of [FIcon]s in a widget subtree.
-class FIconData extends InheritedWidget {
-  /// The data from the closest instance of this class that encloses the given context, if any.
-  static FIconData? maybeOf(BuildContext context) => context.dependOnInheritedWidgetOfExactType<FIconData>();
+class FIconStyle with Diagnosticable {
+  /// The icon style from the closest instance of [FInheritedIconStyle] that encloses the given context, or
+  /// [FStyle.iconStyle] otherwise.
+  static FIconStyle of(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<FInheritedIconStyle>()?.style ?? context.theme.style.iconStyle;
 
   /// The icon's color.
   final Color color;
@@ -16,16 +40,11 @@ class FIconData extends InheritedWidget {
   /// Throws [AssertionError] if `size` is not positive.
   final double size;
 
-  /// Creates a [FIconData].
-  const FIconData({
+  /// Creates a [FIconStyle].
+  const FIconStyle({
     required this.color,
     required this.size,
-    required super.child,
-    super.key,
   }) : assert(0 < size, 'size is $size, but it should be positive.');
-
-  @override
-  bool updateShouldNotify(FIconData old) => color != old.color || size != old.size;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -36,7 +55,15 @@ class FIconData extends InheritedWidget {
   }
 }
 
-/// A graphical icon widget that inherits from a [FIconData], if any.
+/// A graphical icon widget that inherits its style from an [FIconStyle], if any.
+///
+/// [FIconStyle] allows icons to be automatically configured by an enclosing widget. It is explicitly mentioned in
+/// a widget's documentation, such as [FButton.icon], if it provides an [FIconStyle]. [FIcon] defaults to
+/// [FStyle.iconStyle] otherwise.
+///
+/// See:
+/// * [FIconStyle] for the properties that can be inherited.
+/// * [FAssets.icons] for bundled Forui icons.
 abstract class FIcon extends StatelessWidget {
   /// The icon's color.
   final Color? color;
@@ -86,14 +113,34 @@ abstract class FIcon extends StatelessWidget {
 
   /// Creates a [FIcon] from an [ImageProvider].
   ///
+  /// **Note:** Provided images should always have a transparent background. Otherwise, the entire icon will be [color].
+  ///
+  /// Set [color] to [Colors.transparent] to avoid recoloring the image.
+  ///
   /// See [ImageIcon] for more information.
   const factory FIcon.image(
     ImageProvider<Object> image, {
-    Color color,
-    double size,
-    String semanticLabel,
+    Color? color,
+    double? size,
+    String? semanticLabel,
     Key? key,
   }) = _ImageProviderIcon;
+
+  /// Creates a [FIcon] from a [ValueWidgetBuilder].
+  /// 
+  /// To access widget-specific data, i.e. [FButtonData] inside a [FButton]:
+  /// ```dart
+  /// FButton.icon(
+  ///   icon: FIcon.raw(
+  ///     builder: (context, style, _) {
+  ///       final data = FButtonData.of(context);
+  ///       final someWidget = ... // Create someWidget using data here
+  ///       return someWidget;
+  ///     },
+  /// );
+  /// ```
+  const factory FIcon.raw({required ValueWidgetBuilder<FIconStyle> builder, Widget? child, Key? key}) =
+      _BuilderIcon;
 
   const FIcon._({required this.color, this.size, this.semanticLabel, super.key});
 
@@ -134,12 +181,17 @@ class _Icon extends FIcon {
 
   @override
   Widget build(BuildContext context) {
-    final data = FIconData.maybeOf(context);
-    final color = this.color ?? data?.color;
-
+    final data = FIconStyle.of(context);
     return icon.call(
-      colorFilter: color == null ? null : ColorFilter.mode(color, BlendMode.srcIn),
-      height: size ?? data?.size,
+      matchTextDirection: matchTextDirection,
+      fit: fit,
+      alignment: alignment,
+      allowDrawingOutsideViewBox: allowDrawingOutsideViewBox,
+      placeholderBuilder: placeholderBuilder,
+      excludeFromSemantics: excludeFromSemantics,
+      clipBehavior: clipBehavior,
+      colorFilter: ColorFilter.mode(color ?? data.color, BlendMode.srcIn),
+      height: size ?? data.size,
     );
   }
 
@@ -172,7 +224,6 @@ class _IconDataIcon extends FIcon {
   final double? grade;
   final double? opticalSize;
   final List<Shadow>? shadows;
-
   final TextDirection? textDirection;
   final bool applyTextScaling;
 
@@ -193,8 +244,7 @@ class _IconDataIcon extends FIcon {
 
   @override
   Widget build(BuildContext context) {
-    final data = FIconData.maybeOf(context);
-    final color = this.color ?? data?.color;
+    final data = FIconStyle.of(context);
     return Icon(
       icon,
       fill: fill,
@@ -205,8 +255,8 @@ class _IconDataIcon extends FIcon {
       semanticLabel: semanticLabel,
       textDirection: textDirection,
       applyTextScaling: applyTextScaling,
-      color: color,
-      size: size ?? data?.size,
+      color: color ?? data.color,
+      size: size ?? data.size,
     );
   }
 
@@ -239,12 +289,17 @@ class _ImageProviderIcon extends FIcon {
 
   @override
   Widget build(BuildContext context) {
-    final data = FIconData.maybeOf(context);
-    return ImageIcon(
-      image,
-      color: color ?? data?.color,
-      size: size ?? data?.size,
-      semanticLabel: semanticLabel,
+    final data = FIconStyle.of(context);
+    return Semantics(
+      label: semanticLabel,
+      child: Image(
+        image: image,
+        height: size ?? data.size,
+        width: size ?? data.size,
+        color: color == Colors.transparent ? null : (color ?? data.color),
+        fit: BoxFit.scaleDown,
+        excludeFromSemantics: true,
+      ),
     );
   }
 
@@ -252,5 +307,21 @@ class _ImageProviderIcon extends FIcon {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties.add(DiagnosticsProperty('image', image));
+  }
+}
+
+class _BuilderIcon extends FIcon {
+  final ValueWidgetBuilder<FIconStyle> builder;
+  final Widget? child;
+
+  const _BuilderIcon({required this.builder, this.child, super.key}) : super._(color: null, size: null);
+
+  @override
+  Widget build(BuildContext context) => builder(context, FIconStyle.of(context), child);
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(ObjectFlagProperty.has('builder', builder));
   }
 }