+++ forui/test/src/widgets/alert_golden_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:forui/forui.dart';
-import 'package:forui/src/widgets/alert/alert.dart';
-import '../../test_scaffold.dart';
+import 'package:forui/src/widgets/alert.dart';
+import '../test_scaffold.dart';
 
 void main() {
   group('FAlert', () {