+++ forui/test/src/widgets/bottom_navigation_bar_golden_test.dart
@@ -20,23 +20,23 @@ void main() {
               index: 2,
               children: [
                 FBottomNavigationBarItem(
-                  icon: FAssets.icons.home,
+                  icon: FIcon(FAssets.icons.home),
                   label: const Text('Home'),
                 ),
                 FBottomNavigationBarItem(
-                  icon: FAssets.icons.layoutGrid,
+                  icon: FIcon(FAssets.icons.layoutGrid),
                   label: const Text('Browse'),
                 ),
                 FBottomNavigationBarItem(
-                  icon: FAssets.icons.radio,
+                  icon: FIcon(FAssets.icons.radio),
                   label: const Text('Radio'),
                 ),
                 FBottomNavigationBarItem(
-                  icon: FAssets.icons.libraryBig,
+                  icon: FIcon(FAssets.icons.radio),
                   label: const Text('Library'),
                 ),
                 FBottomNavigationBarItem(
-                  icon: FAssets.icons.search,
+                  icon: FIcon(FAssets.icons.radio),
                   label: const Text('Search'),
                 ),
               ],
@@ -47,51 +47,6 @@ void main() {
         await expectLater(
           find.byType(TestScaffold),
           matchesGoldenFile('bottom-navigation-bar/forui-icon-$name.png'),
-        );
-      });
-
-      testWidgets('custom icon - $name', (tester) async {
-        Widget icon(BuildContext _, FBottomNavigationBarData data, Widget? __) => Container(
-              height: data.itemStyle.iconSize,
-              width: data.itemStyle.iconSize,
-              color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
-            );
-
-        await tester.pumpWidget(
-          TestScaffold(
-            data: theme,
-            background: background,
-            child: FBottomNavigationBar(
-              index: 2,
-              children: [
-                FBottomNavigationBarItem.custom(
-                  iconBuilder: icon,
-                  label: const Text('Home'),
-                ),
-                FBottomNavigationBarItem.custom(
-                  iconBuilder: icon,
-                  label: const Text('Browse'),
-                ),
-                FBottomNavigationBarItem.custom(
-                  iconBuilder: icon,
-                  label: const Text('Radio'),
-                ),
-                FBottomNavigationBarItem.custom(
-                  iconBuilder: icon,
-                  label: const Text('Library'),
-                ),
-                FBottomNavigationBarItem.custom(
-                  iconBuilder: icon,
-                  label: const Text('Search'),
-                ),
-              ],
-            ),
-          ),
-        );
-
-        await expectLater(
-          find.byType(TestScaffold),
-          matchesGoldenFile('bottom-navigation-bar/custom-icon-$name.png'),
         );
       });
     }