+++ forui/CHANGELOG.md
@@ -3,8 +3,6 @@
 ### Additions
 * Add `FSlider`
 
-* Add `FBottomNavigationBarItem.custom(...)`.
-
 * Add `FButtonCustomStyle.enabledHoverBoxDecoration`.
 
 * Add `FTextField.contentInsertionConfiguration`.
@@ -45,6 +43,10 @@
 
 * * **Breaking** Change FButtonCustomStyle to better represent the style's layout - this will only affect users that 
   create a custom `FButtonCustomStyle`.
+
+* **Breaking** Change `FBottomNavigationBarItem.icon` from `SvgAsset` to `Widget` - wrap the asset in ` FIcon` instead.
+
+* **Breaking** Change `FHeaderAction.icon` from `SvgAsset` to `Widget` - wrap the asset in ` FIcon` instead.
 
 ### Fixes